reorganized code to create a baseline
