--- app/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "code-engine-gallery",
-  "version": "1.0.0",
-  "description": "Sample app to demonstrate power of Code Engine components",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "ibm-cos-sdk": "^1.13.1"
-  }
-}
