--- build
@@ -1,25 +1,0 @@
-#!/bin/bash
-
-# Env Vars:
-# REGISTRY: name of the image registry/namespace to store the images
-# NOCACHE: set this to "--no-cache" to turn off the Docker build cache
-#
-# NOTE: to run this you MUST set the REGISTRY environment variable to
-# your own image registry/namespace otherwise the `docker push` commands
-# will fail due to an auth failure. Which means, you also need to be logged
-# into that registry before you run it.
-
-set -ex
-export REGISTRY=${REGISTRY:-sathvik157/sathvik-image-app}
-
-# First build and push the app - notice no tag on the image name
-cd app
-podman build ${NOCACHE} -t ${REGISTRY}/gallery . --platform linux/amd64
-podman push ${REGISTRY}/gallery:latest
-cd ..
-
-# Second build and push the job - notice no tag on the image name
-cd job
-podman build ${NOCACHE} -t ${REGISTRY}/gallery-job . --platform linux/amd64
-podman push ${REGISTRY}/gallery-job:latest
-cd ..
