+++ colorizer.js
