+++ cos-service.js
