--- job/Dockerfile
@@ -1,18 +1,0 @@
-FROM ubuntu:latest
-
-# Install Node.js dependencies
-RUN apt-get update && \
-    apt-get install -y curl && \
-    curl -fsSL https://deb.nodesource.com/setup_18.x | bash - && \
-    apt-get install -y nodejs
-
-# Define which files should be copied into the container image
-COPY *.js .
-COPY package.json .
-
-# Load all dependencies
-RUN npm install
-
-# Define how the job is being started
-CMD [ "node", "job.js" ]
-
